+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Master
 
+## v0.1.0
+
 * Highlight source in console when warning conversion error
 * Add `--no-color` option